Added some comments

+++ calculator.py
@@ -1,4 +1,11 @@
-students = int(input("Enter number of students: "))
-seats = int(input("Enter number of seats: "))
+#conversion or type casting
+
+
+#Implicit Conversion (Type Coercion)
+#Explicit Conversion (Type Casting)
+
+#means that makes the code more readable#
+students = float(input("Enter number of students: "))
+seats = float(input("Enter number of seats: "))
 
 print(students / seats)
